requirements.txt
flask
requests
python-dotenv
openai
pydantic
pydantic-settings
Pillow